refactor acc opponents hud config
