+++ Race_Element.HUD.ACC/Overlays/Driving/Opponents/OpponentsOverlay.cs
@@ -15,27 +15,7 @@ Description = "Shows information about the cars ahead and behind in terms of rac
 internal sealed class OpponentsOverlay : AbstractOverlay
 {
     private readonly OpponentsConfiguration _config = new();
-    private sealed class OpponentsConfiguration : OverlayConfiguration
-    {
-        [ConfigGrouping("Opponents", "Change how many opponents are displayed either ahead or behind")]
-        public OpponentsGrouping Opponents { get; init; } = new();
-        public sealed class OpponentsGrouping
-        {
-            [IntRange(1, 5, 1)]
-            public int AheadCount { get; init; } = 1;
-            [IntRange(1, 5, 1)]
-            public int BehindCount { get; init; } = 1;
-        }
-
-        [ConfigGrouping("Data", "Change which opponents data is displayed.")]
-        public BehaviorGrouping Data { get; init; } = new();
-        public sealed class BehaviorGrouping
-        {
-            public bool Sectors { get; init; } = true;
-            public bool Gap { get; init; } = true;
-            public bool Difference { get; init; } = true;
-        }
-    }
+  
     private readonly record struct OpponentsModel
     {
         public CarDataModel[] Ahead { get; init; }
@@ -62,7 +42,6 @@ internal sealed class OpponentsOverlay : AbstractOverlay
         List<int> columnSizes = [50];
 
         _table = new InfoTable(12, [100, 100, 50]) { DrawBackground = true, DrawRowLines = true, DrawValueBackground = true, };
-
     }
 
     public sealed override bool ShouldRender() => true;
